Fast/requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.8
flatbuffers==25.12.19
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
narwhals==2.17.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.19.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==4.25.8
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2026.1.post1
referencing==0.36.2
requests==2.32.5
rich==14.3.3
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
starlette==0.49.3
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
uvicorn==0.39.0
watchdog==6.0.0
Werkzeug==3.1.6
wrapt==2.1.2
zipp==3.23.0
langchain
langchain-groq
